Document migrations command
Briefly explain why the migration check was moved to a CLI command
and provide instructions for execution.

In hindsight, a better solution may have been to leave the checks
where they were but throw a warning rather than an exception. I still
believe that the check should be done during the "execution" of the
code and not the "compilation".